refactor: separate emoji picker into files
